fix error

+++ database/migrations/2023_05_16_063731_create_products_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->int('user_id');
-            $table->int('category_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('category_id');
             $table->string('name');
             $table->text('description');
             $table->bigInteger('starting_price');
